requirements.txt
python-dotenv
discord.py==2.0.0
langchain
langchain-google-vertexai
anthropic[vertex]
